+++ core/AppSpec.ts
@@ -52,8 +52,15 @@ export class AppSpec<T extends HoistAppModel = HoistAppModel> {
     /** True if the app should use the Hoist mobile toolkit.*/
     isMobileApp: boolean;
 
-    /** True if SSO auth is enabled, as opposed to a login prompt. */
-    isSSO: boolean;
+    /** True to show a login form on initialization when not authenticated. (default false) */
+    enableLoginForm?: boolean;
+
+    /**
+     * True to show logout options in the app.
+     *
+     * For apps with auth schemes that can support this operation (e.g. OAuth).  (default false)
+     */
+    enableLogout?: boolean;
 
     /**
      * Method for determining if user may access the app.
@@ -86,7 +93,7 @@ export class AppSpec<T extends HoistAppModel = HoistAppModel> {
      */
     lockoutPanel?: ElementFactory | FunctionComponent | Component;
 
-    /** Optional message to show on login form (for non-SSO apps). */
+    /** Optional message to show on login form (see showLoginForm). */
     loginMessage?: string;
 
     /** Optional message to show users when denied access to app. */
@@ -114,8 +121,9 @@ export class AppSpec<T extends HoistAppModel = HoistAppModel> {
         componentClass,
         containerClass,
         isMobileApp,
-        isSSO,
         checkAccess,
+        enableLoginForm = false,
+        enableLogout = false,
         trackAppLoad = true,
         webSocketsEnabled = false,
         idlePanel = null,
@@ -134,7 +142,6 @@ export class AppSpec<T extends HoistAppModel = HoistAppModel> {
                 isMobileApp ? 'mobile' : 'desktop'
             }/AppContainer".`
         );
-        throwIf(isNil(isSSO), 'A Hoist App must define isSSO');
 
         throwIf(
             !isString(checkAccess) && !isFunction(checkAccess),
@@ -148,13 +155,14 @@ export class AppSpec<T extends HoistAppModel = HoistAppModel> {
         this.componentClass = componentClass;
         this.containerClass = containerClass;
         this.isMobileApp = isMobileApp;
-        this.isSSO = isSSO;
         this.checkAccess = checkAccess;
 
         this.trackAppLoad = trackAppLoad;
         this.webSocketsEnabled = webSocketsEnabled;
         this.idlePanel = idlePanel;
         this.lockoutPanel = lockoutPanel;
+        this.enableLogout = enableLogout;
+        this.enableLoginForm = enableLoginForm;
         this.loginMessage = loginMessage;
         this.lockoutMessage = lockoutMessage;
         this.showBrowserContextMenu = showBrowserContextMenu;