+++ CHANGELOG.md
@@ -49,6 +49,11 @@ for more details.
 * Updated CSS variables related to the `ZoneGrid` component - vars formerly prefixed
   by `--xh-grid-multifield` are now prefixed by `--xh-zone-grid`, several vars have been added, and
   some defaults have changed.
+* Removed obsolete `AppSpec.isSSO` property in favor of two new properties `AppSpec.enableLogout`
+  and `AppSpec.enableLoginForm`.  This should have no effect on the vast majority of apps which had
+  `isSSO` set to `true`.  For apps where `isSSO` was set false, the new flags should be
+  used to more clearly indicate the desired auth behavior.
+
 
 ### 🎁 New Features
 
@@ -61,8 +66,10 @@ for more details.
 * Improved the display of Role details in the Admin Console. The detail panel for the selected role
   now includes a sub-tab listing all other roles inherited by the selected role, something that
   was previously accessible only via the linked graph visualization.
-* Added new `checkboxRenderer` for rendering booleans with a checkbox input look and feel.
-* Added new mobile `checkboxButton`, an alternate input component for toggling boolean values.
+* Added new `checkBoxRenderer` for rendering booleans with a checkbox input look and feel.
+* Added beta version of a new Hoist `security` package, providing built-in support for OAuth flows.
+  See `BaseOAuthClient`, `MsalClient`, and `AuthZeroClient` for more information.  Please note that
+  package is being released as a *beta* and is subject to change before final release.
 
 ### ✨ Styles
 
@@ -91,11 +98,16 @@ for more details.
 
 * NumberEditor no longer activates on keypress of letter characters.
 * Removed initial `ping` call `FetchService` init.
+* Deprecated `FetchService.setDefaultHeaders` and replaced with new `addDefaultHeaders` method to
+  support independent additions of default headers from multiple sources in an application.
 
 ### 📚 Libraries
 
 * @ag-grid `30.x → 31.x`
+* @auth0/auth0-spa-js `added @ 2.1`
+* @azure/msal-browser `added @ 3.14`
 * dompurify `3.0 → 3.1`
+* jwt-decode `added @ 4.0`
 * moment `2.29 → 2.30`
 * numbro `2.4 → 2.5`
 * qs `6.11 → 6.12`